% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shatterseek.R
\docType{class}
\name{CNVsegs-class}
\alias{CNVsegs-class}
\alias{CNVsegs}
\title{Class to store CNV data}
\arguments{
\item{chrom}{(character): chromosome (also in Ensembl notation)}

\item{start}{(numeric): start position for the CN segment}

\item{end}{(numeric): end position for the CN segment}

\item{CN}{(numeric): integer total copy number (e.g. 2 for unaltered chromosomal regions)}
}
\value{
an instance of the class 'CNVsegs' that contains CNV data. Required format by the function shatterseek
}
\description{
Class to store CNV data
}
