\name{plot.chromoth}

\alias{plot.chromoth}



\title{Plot a \link{chromoth} object given by \link{shatterseek}}

\description{
	Plot the copy number profile and structural variations of a \link{chromoth} object given by \link{shatterseek}.
	}

\usage{ 
        plot.chromoth(chromothripsis,chrom,sample)
        } 
\arguments{
	\item{chromothripsis}{
		A \link{chromoth} object given by \link{shatterseek}. 
		}
	\item{chrom.name}{
		A character string. The name of the chromosome to be plotted.
		}
	\item{sample.name}{
		A character string. The name of the sample. This is optional.
		}
	}

\details{
	
	}


\value{
	}

\seealso{
	\link{chromoth},\link{shatterseek}
	}

\author{Ruibin Xi}
