% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shatterseek.R
\docType{class}
\name{SVs-class}
\alias{SVs-class}
\alias{SVs}
\title{Class to store SV data}
\arguments{
\item{chrom1}{(character): chromosome for the first breakpoint}

\item{pos1}{(character): position for the first breakpoint}

\item{chrom2}{(character): chromosome for the second breakpoint}

\item{pos2}{(character): position for the second breakpoint}

\item{SVtype}{(character): type of SV, encoded as: DEL (deletion-like; +/-), DUP (duplication-like; -/+), h2hINV (head-to-head inversion; +/+), and t2tINV (tail-to-tail inversion; -/-).}

\item{strand1}{(e.g. + for DEL)}

\item{strand2}{(e.g. - for DEL)}
}
\value{
an instance of the class 'SVs' that contains SV data. Required format by the function shatterseek
}
\description{
Class to store SV data
}
